requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Faker==37.1.0
frozenlist==1.6.0
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pyTelegramBotAPI==4.26.0
python-decouple==3.8
python-telegram-bot==22.0
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.40
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
python-decouple